pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apm_crewconnect"
version = "0.0.10"
authors = [
  { name="Hugo Clarke-Wing", email="hugo@clarkewing.com" },
]
description = "A package to interact with the APM CrewConnect API"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "pyhumps>=3.8.0",
  "requests_oauthlib>=2.0.0",
  "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/clarkewing/apm_crewconnect"
Issues = "https://github.com/clarkewing/apm_crewconnect/issues"
